+++ CoreOpenGLDisplay/src/au/gov/asd/tac/constellation/visual/opengl/utilities/glyphs/GlyphsFrame.java
@@ -409,7 +409,7 @@ public class GlyphsFrame extends JFrame {
 
     private static String[] loadText(final String fnam, final boolean raw) throws IOException {
         try (final BufferedReader in = new BufferedReader(new InputStreamReader(GlyphsFrame.class.getResourceAsStream(fnam), StandardCharsets.UTF_8))) {
-            final List<String> ls = in.lines().filter(line -> raw || (line.length() > 0 && !line.startsWith("#"))).toList();
+            final List<String> ls = in.lines().filter(line -> raw || (!line.isEmpty() && !line.startsWith("#"))).toList();
             return ls.toArray(new String[ls.size()]);
         }
     }