+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Changes in October 2025
 -   Removed `areSame()` from `BoundingBox`, `Camera`, `Frame`, `Vector3f` in favour of implementing `equals()`.
+-   Removed `Graphics3DUtilities.clamp()` which is no longer required now that `Math.clamp()` does the same thing.
 
 ## Changes in September 2025
 -   Removed `DefaultGetter` and `PrimaryKeyDefaultGetter` due to limited implementations. Uses now instead will use default value directly. e.g. Constructors and `createEditor()` now pass the default value instead of the `DefaultGetter`